fix warnings

See merge request integer/papilo!108

+++ src/papilo/core/ProblemBuilder.hpp
@@ -111,7 +111,7 @@ class ProblemBuilder
    setObjAll( Vec<REAL> values )
    {
       assert( values.size() == obj.coefficients.size() );
-      for( int c = 0; c < values.size(); ++c )
+      for( int c = 0; c < (int) values.size(); ++c )
          obj.coefficients[c] = std::move( values[c] );
    }
 
@@ -135,7 +135,7 @@ class ProblemBuilder
    setColLbInfAll( Vec<uint8_t> isInfinite )
    {
       assert( domains.flags.size() == isInfinite.size() );
-      for( int c = 0; c < isInfinite.size(); ++c )
+      for( int c = 0; c < (int) isInfinite.size(); ++c )
          setColLbInf( c, isInfinite[c] );
    }
 
@@ -152,7 +152,7 @@ class ProblemBuilder
    setColUbInfAll( Vec<uint8_t> isInfinite )
    {
       assert( domains.flags.size() == isInfinite.size() );
-      for( int c = 0; c < isInfinite.size(); ++c )
+      for( int c = 0; c < (int) isInfinite.size(); ++c )
          setColUbInf( c, isInfinite[c] );
    }
 
@@ -166,8 +166,9 @@ class ProblemBuilder
    setColLbAll( Vec<REAL> lbs )
    {
       assert( lbs.size() == domains.lower_bounds.size() );
-      for( int c = 0; c < lbs.size(); ++c )
-         domains.lower_bounds[c] = std::move( lbs[c] );
+      for( int c = 0; c < (int) lbs.size(); ++c )
+         domains.lower_bounds[c] = std::move( lbs[c
+         ] );
    }
 
    void
@@ -180,7 +181,7 @@ class ProblemBuilder
    setColUbAll( Vec<REAL> ubs )
    {
       assert( ubs.size() == domains.upper_bounds.size() );
-      for( int c = 0; c < ubs.size(); ++c )
+      for( int c = 0; c < (int) ubs.size(); ++c )
          domains.upper_bounds[c] = std::move( ubs[c] );
    }
 
@@ -197,7 +198,7 @@ class ProblemBuilder
    setColIntegralAll( Vec<uint8_t> isIntegral )
    {
       assert( isIntegral.size() == domains.flags.size() );
-      for( int c = 0; c < isIntegral.size(); ++c )
+      for( int c = 0; c < (int) isIntegral.size(); ++c )
          setColIntegral( c, isIntegral[c] );
    }
 
@@ -214,7 +215,7 @@ class ProblemBuilder
    setRowLhsInfAll( Vec<uint8_t> isInfinite )
    {
       assert( isInfinite.size() == rflags.size() );
-      for( int r = 0; r < isInfinite.size(); ++r )
+      for( int r = 0; r < (int) isInfinite.size(); ++r )
          setRowLhsInf( r, isInfinite[r] );
    }
 
@@ -231,7 +232,7 @@ class ProblemBuilder
    setRowRhsInfAll( Vec<uint8_t> isInfinite )
    {
       assert( isInfinite.size() == rflags.size() );
-      for( int r = 0; r < isInfinite.size(); ++r )
+      for( int r = 0; r < (int) isInfinite.size(); ++r )
          setRowRhsInf( r, isInfinite[r] );
    }
 
@@ -245,7 +246,7 @@ class ProblemBuilder
    setRowLhsAll( Vec<REAL> lhsvals )
    {
       assert( lhsvals.size() == lhs.size() );
-      for( int r = 0; r < lhsvals.size(); ++r )
+      for( int r = 0; r < (int) lhsvals.size(); ++r )
          lhs[r] = std::move( lhsvals[r] );
    }
 
@@ -259,7 +260,7 @@ class ProblemBuilder
    setRowRhsAll( Vec<REAL> rhsvals )
    {
       assert( rhsvals.size() == rhs.size() );
-      for( int r = 0; r < rhsvals.size(); ++r )
+      for( int r = 0; r < (int) rhsvals.size(); ++r )
          rhs[r] = std::move( rhsvals[r] );
    }
 
@@ -303,7 +304,7 @@ class ProblemBuilder
    setRowNameAll( Vec<Str> names )
    {
       assert( rownames.size() == names.size() );
-      for( int r = 0; r < names.size(); ++r )
+      for( int r = 0; r < (int) names.size(); ++r )
          rownames[r] = String( names[r] );
    }
 
@@ -319,7 +320,7 @@ class ProblemBuilder
    setColNameAll( Vec<Str> names )
    {
       assert( colnames.size() == names.size() );
-      for( int c = 0; c < names.size(); ++c )
+      for( int c = 0; c < (int) names.size(); ++c )
          colnames[c] = String( names[c] );
    }
